+++ mapsforge-samples-android/src/main/java/org/mapsforge/samples/android/ColorFilterMapViewer.java
@@ -18,6 +18,8 @@ import android.graphics.*;
 import org.mapsforge.core.graphics.Bitmap;
 import org.mapsforge.map.android.graphics.AndroidBitmap;
 import org.mapsforge.map.android.graphics.AndroidGraphicFactory;
+import org.mapsforge.map.datastore.PointOfInterest;
+import org.mapsforge.map.layer.renderer.PolylineContainer;
 import org.mapsforge.map.rendertheme.ThemeCallback;
 import org.mapsforge.map.rendertheme.renderinstruction.RenderInstruction;
 
@@ -56,6 +58,16 @@ public class ColorFilterMapViewer extends DefaultTheme {
                 // Grayscale
                 r = g = b = (int) (0.213f * r + 0.715f * g + 0.072f * b);
                 return (a << 24) | (r << 16) | (g << 8) | b;
+            }
+
+            @Override
+            public String getText(PointOfInterest poi, String text) {
+                return text;
+            }
+
+            @Override
+            public String getText(PolylineContainer way, String text) {
+                return text;
             }
         });
     }