+++ docs/Changelog.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+- Motorider map theme improvements [#1483](https://github.com/mapsforge/mapsforge/issues/1483)
+- Biker map theme improvements [#1674](https://github.com/mapsforge/mapsforge/issues/1674)
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.26.0)