+++ docs/Changelog.md
@@ -2,6 +2,11 @@
 
 ## Next version
 
+- Minor improvements and bug fixes
+- [Solved issues](https://github.com/mapsforge/mapsforge/issues?q=is%3Aclosed+milestone%3A0.27.0)
+
+## Version 0.26.0 (2025-09-14)
+
 - MBTiles raster tile source (Android) [#1687](https://github.com/mapsforge/mapsforge/pull/1687)
 - Render themes: `Display.FORCED` parameter [#1717](https://github.com/mapsforge/mapsforge/pull/1717)
 - `ThemeCallback.getText` method [#1700](https://github.com/mapsforge/mapsforge/pull/1700)